Change type of data_size_t to int64_t

+++ include/LightGBM/c_api.h
@@ -300,7 +300,7 @@ LIGHTGBM_C_EXPORT int LGBM_DatasetCreateFromMats(int32_t nmat,
  * \return 0 when succeed, -1 when failure happens
  */
 LIGHTGBM_C_EXPORT int LGBM_DatasetGetSubset(const DatasetHandle handle,
-                                            const int32_t* used_row_indices,
+                                            const int64_t *used_row_indices,
                                             int32_t num_used_row_indices,
                                             const char* parameters,
                                             DatasetHandle* out);
